move payment API client methods into a single file
