--- internal/sirius/add_payment_test.go
@@ -1,69 +1,0 @@
-package sirius
-
-import (
-	"context"
-	"fmt"
-	"github.com/pact-foundation/pact-go/v2/consumer"
-	"github.com/pact-foundation/pact-go/v2/matchers"
-	"github.com/stretchr/testify/assert"
-	"net/http"
-	"testing"
-)
-
-func TestAddPayment(t *testing.T) {
-	t.Parallel()
-
-	pact, err := newPact()
-	assert.NoError(t, err)
-
-	testCases := []struct {
-		name          string
-		setup         func()
-		expectedError func(int) error
-	}{
-		{
-			name: "OK",
-			setup: func() {
-				pact.
-					AddInteraction().
-					Given("I have a pending case assigned").
-					UponReceiving("A request to create a payment").
-					WithCompleteRequest(consumer.Request{
-						Method: http.MethodPost,
-						Path:   matchers.String("/lpa-api/v1/cases/800/payments"),
-						Headers: matchers.MapMatcher{
-							"Content-Type": matchers.String("application/json"),
-						},
-						Body: map[string]interface{}{
-							"amount":      4100,
-							"source":      "PHONE",
-							"paymentDate": "25/04/2022",
-						},
-					}).
-					WithCompleteResponse(consumer.Response{
-						Status:  http.StatusCreated,
-						Headers: matchers.MapMatcher{"Content-Type": matchers.String("application/json")},
-					})
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			tc.setup()
-
-			assert.Nil(t, pact.ExecuteTest(t, func(config consumer.MockServerConfig) error {
-				client := NewClient(http.DefaultClient, fmt.Sprintf("http://127.0.0.1:%d", config.Port))
-
-				err := client.AddPayment(Context{Context: context.Background()}, 800, 4100, "PHONE", DateString("2022-04-25"))
-
-				if tc.expectedError == nil {
-					assert.Nil(t, err)
-				} else {
-					assert.Equal(t, tc.expectedError(config.Port), err)
-				}
-				return nil
-			}))
-		})
-	}
-}
