--- internal/sirius/delete_payment.go
@@ -1,9 +1,0 @@
-package sirius
-
-import (
-	"fmt"
-)
-
-func (c *Client) DeletePayment(ctx Context, paymentID int) error {
-	return c.delete(ctx, fmt.Sprintf("/lpa-api/v1/payments/%d", paymentID))
-}
