--- internal/sirius/delete_payment_test.go
@@ -1,109 +1,0 @@
-package sirius
-
-import (
-	"context"
-	"fmt"
-	"github.com/pact-foundation/pact-go/v2/consumer"
-	"github.com/pact-foundation/pact-go/v2/matchers"
-	"github.com/stretchr/testify/assert"
-	"net/http"
-	"testing"
-)
-
-func TestDeletePayment(t *testing.T) {
-	t.Parallel()
-
-	pact, err := newPact()
-	assert.NoError(t, err)
-
-	testCases := []struct {
-		name          string
-		setup         func()
-		expectedError func(int) error
-	}{
-		{
-			name: "OK",
-			setup: func() {
-				pact.
-					AddInteraction().
-					Given("I have an lpa which has been paid for").
-					UponReceiving("A request to delete a payment").
-					WithCompleteRequest(consumer.Request{
-						Method: http.MethodDelete,
-						Path:   matchers.String("/lpa-api/v1/payments/123"),
-					}).
-					WithCompleteResponse(consumer.Response{
-						Status: http.StatusNoContent,
-					})
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			tc.setup()
-
-			assert.Nil(t, pact.ExecuteTest(t, func(config consumer.MockServerConfig) error {
-				client := NewClient(http.DefaultClient, fmt.Sprintf("http://127.0.0.1:%d", config.Port))
-
-				err := client.DeletePayment(Context{Context: context.Background()}, 123)
-
-				if tc.expectedError == nil {
-					assert.Nil(t, err)
-				} else {
-					assert.Equal(t, tc.expectedError(config.Port), err)
-				}
-				return nil
-			}))
-		})
-	}
-}
-
-func TestDeleteFeeReduction(t *testing.T) {
-	t.Parallel()
-
-	pact, err := newPact()
-	assert.NoError(t, err)
-
-	testCases := []struct {
-		name          string
-		setup         func()
-		expectedError func(int) error
-	}{
-		{
-			name: "OK",
-			setup: func() {
-				pact.
-					AddInteraction().
-					Given("I have an lpa which has a fee reduction").
-					UponReceiving("A request to delete a fee reduction").
-					WithCompleteRequest(consumer.Request{
-						Method: http.MethodDelete,
-						Path:   matchers.String("/lpa-api/v1/payments/124"),
-					}).
-					WithCompleteResponse(consumer.Response{
-						Status: http.StatusNoContent,
-					})
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			tc.setup()
-
-			assert.Nil(t, pact.ExecuteTest(t, func(config consumer.MockServerConfig) error {
-				client := NewClient(http.DefaultClient, fmt.Sprintf("http://127.0.0.1:%d", config.Port))
-
-				err := client.DeletePayment(Context{Context: context.Background()}, 124)
-
-				if tc.expectedError == nil {
-					assert.Nil(t, err)
-				} else {
-					assert.Equal(t, tc.expectedError(config.Port), err)
-				}
-				return nil
-			}))
-		})
-	}
-}
