+++ web/template/layout/payment.gohtml
@@ -22,9 +22,7 @@
                     <tr class="govuk-table__row">
                         <th scope="row" class="govuk-table__header">Method:</th>
                         <td class="govuk-table__cell">
-                            source: {{ $p.Source }}
                             {{ range $.PaymentSources }}
-                                handle: {{ .Handle }}
                                 {{ if eq .Handle $p.Source }}{{ .Label }}{{ end }}
                             {{ end }}
                         </td>
